// web/sdk/src/models/SiteConfigApiOut.ts
export interface SiteConfigApiOut {
    /**
     * 
     * @type {string}
     * @memberof SiteConfigApiOut
     */
    name: string;
    /**
     * 
     * @type {string}
     * @memberof SiteConfigApiOut
     */
    siteUrl: string;
    /**
     * 
     * @type {string}
     * @memberof SiteConfigApiOut
     */
    ownerUserId?: string | null;
    /**
     * 
     * @type {string}
     * @memberof SiteConfigApiOut
     */
    successTextClass?: string;
    /**
     * 
     * @type {string}
     * @memberof SiteConfigApiOut
     */
    expectedSuccessText?: string;
    /**
     * 
     * @type {Array<string>}
     * @memberof SiteConfigApiOut
     */
    requiredCookies?: Array<string>;
    /**
     * 
     * @type {string}
     * @memberof SiteConfigApiOut
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof SiteConfigApiOut
     */
    loginType?: SiteConfigApiOutLoginTypeEnum;
    /**
     * 
     * @type {ApiConfig}
     * @memberof SiteConfigApiOut
     */
    apiConfig: ApiConfig;
}


/**
 * @export
 */
export const SiteConfigApiOutLoginTypeEnum = {
    Api: 'api'
} as const;
export type SiteConfigApiOutLoginTypeEnum = typeof SiteConfigApiOutLoginTypeEnum[keyof typeof SiteConfigApiOutLoginTypeEnum];