// web/sdk/src/models/SiteConfigSeleniumOut.ts
/* tslint:disable */
/* eslint-disable */
/**
 * SubPaperFlux API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { SeleniumConfig } from './SeleniumConfig';
import {
    SeleniumConfigFromJSON,
    SeleniumConfigFromJSONTyped,
    SeleniumConfigToJSON,
    SeleniumConfigToJSONTyped,
} from './SeleniumConfig';

/**
 * 
 * @export
 * @interface SiteConfigSeleniumOut
 */
export interface SiteConfigSeleniumOut {
    /**
     * 
     * @type {string}
     * @memberof SiteConfigSeleniumOut
     */
    name: string;
    /**
     * 
     * @type {string}
     * @memberof SiteConfigSeleniumOut
     */
    siteUrl: string;
    /**
     * 
     * @type {string}
     * @memberof SiteConfigSeleniumOut
     */
    ownerUserId?: string | null;
    /**
     * 
     * @type {string}
     * @memberof SiteConfigSeleniumOut
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof SiteConfigSeleniumOut
     */
    loginType?: SiteConfigSeleniumOutLoginTypeEnum;
    /**
     * 
     * @type {SeleniumConfig}
     * @memberof SiteConfigSeleniumOut
     */
    seleniumConfig: SeleniumConfig;
}


/**
 * @export
 */
export const SiteConfigSeleniumOutLoginTypeEnum = {
    Selenium: 'selenium'
} as const;
export type SiteConfigSeleniumOutLoginTypeEnum = typeof SiteConfigSeleniumOutLoginTypeEnum[keyof typeof SiteConfigSeleniumOutLoginTypeEnum];


/**
 * Check if a given object implements the SiteConfigSeleniumOut interface.
 */
export function instanceOfSiteConfigSeleniumOut(value: object): value is SiteConfigSeleniumOut {
    if (!('name' in value) || value['name'] === undefined) return false;
    if (!('siteUrl' in value) || value['siteUrl'] === undefined) return false;
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('seleniumConfig' in value) || value['seleniumConfig'] === undefined) return false;
    return true;
}

export function SiteConfigSeleniumOutFromJSON(json: any): SiteConfigSeleniumOut {
    return SiteConfigSeleniumOutFromJSONTyped(json, false);
}

export function SiteConfigSeleniumOutFromJSONTyped(json: any, ignoreDiscriminator: boolean): SiteConfigSeleniumOut {
    if (json == null) {
        return json;
    }
    return {
        
        'name': json['name'],
        'siteUrl': json['site_url'],
        'ownerUserId': json['owner_user_id'] == null ? undefined : json['owner_user_id'],
        'id': json['id'],
        'loginType': json['login_type'] == null ? undefined : json['login_type'],
        'seleniumConfig': SeleniumConfigFromJSON(json['selenium_config']),
    };
}

export function SiteConfigSeleniumOutToJSON(json: any): SiteConfigSeleniumOut {
    return SiteConfigSeleniumOutToJSONTyped(json, false);
}

export function SiteConfigSeleniumOutToJSONTyped(value?: SiteConfigSeleniumOut | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'name': value['name'],
        'site_url': value['siteUrl'],
        'owner_user_id': value['ownerUserId'],
        'id': value['id'],
        'login_type': value['loginType'],
        'selenium_config': SeleniumConfigToJSON(value['seleniumConfig']),
    };
}